Improved number parsing as well as tuples

+++ Tokens/NumberToken.cs
@@ -3,23 +3,23 @@ using System;
 [System.Diagnostics.DebuggerDisplay("{Location} {Kind} : {val}")]
 public class NumberToken : ComplexToken
 {
-    public new static readonly NumberToken NULL = new NumberToken(0M, LocationRange.NULL, false);
+    public new static readonly NumberToken NULL = new NumberToken(0d, LocationRange.NULL, false);
 
-    protected decimal val;
+    protected double val;
 
-    public decimal Value { get => val; }
+    public double Value { get => val; }
 
     public NumberToken(string representation, LocationRange location, bool isValid = true, TriviaToken? leading = null, TriviaToken? trailing = null)
         : base(representation, TokenKind.number, location, isValid, leading, trailing)
     {
-        if (isValid && representation.Length != 0 && !Decimal.TryParse(representation, out val))
+        if (isValid && representation.Length != 0 && !Double.TryParse(representation, out val))
             throw Logger.Fatal(new InternalErrorException(
                 message: "This NumberToken has been marked valid, but could not parse string '" + representation + "' as a number",
                 range: Location
             ));
     }
 
-    public NumberToken(decimal d, LocationRange location, bool isValid = true, TriviaToken? leading = null, TriviaToken? trailing = null)
+    public NumberToken(double d, LocationRange location, bool isValid = true, TriviaToken? leading = null, TriviaToken? trailing = null)
         : base(d.ToString().ToLower(), TokenKind.number, location, isValid, leading, trailing)
     {
         val = d;
@@ -30,7 +30,7 @@ public class NumberToken : ComplexToken
     // However, in the state they are right now, they cannot really be used to create a number, so meh.
     public override void Add(char ch) {
         base.Add(ch);
-        if (IsValid && !Decimal.TryParse(Representation, out val))
+        if (IsValid && !Double.TryParse(Representation, out val))
             throw Logger.Fatal(new InternalErrorException(
                 message: "This NumberToken has been marked valid, but could not parse string '" + Representation + "' as a number",
                 range: Location
@@ -39,7 +39,7 @@ public class NumberToken : ComplexToken
 
     public override void Add(string str) {
         base.Add(str);
-        if (IsValid && !Decimal.TryParse(Representation, out val))
+        if (IsValid && !Double.TryParse(Representation, out val))
             throw Logger.Fatal(new InternalErrorException(
                 message: "This NumberToken has been marked valid, but could not parse string '" + Representation + "' as a number",
                 range: Location
